Remove util.js
