+++ modules/shapefile/src/lib/parse-shp.js
@@ -1,6 +1,7 @@
-import {parseHeader, BIG_ENDIAN, LITTLE_ENDIAN} from './util';
 import {parseRecord} from './parse-geometry';
 
+const LITTLE_ENDIAN = true;
+const BIG_ENDIAN = false;
 const SHAPE_HEADER_SIZE = 100;
 // According to the spec, the record header is just 8 bytes, but here we set it
 // to 12 so that we can also access the record's type
@@ -43,5 +44,23 @@ export default function parseShape(arrayBuffer) {
   return {
     header,
     features
+  };
+}
+
+export function parseHeader(header) {
+  return {
+    // Length is stored as # of 2-byte words; multiply by 2 to get # of bytes
+    length: header.getInt32(24, BIG_ENDIAN) * 2,
+    type: header.getInt32(32, LITTLE_ENDIAN),
+    bbox: {
+      minX: header.getFloat64(36, LITTLE_ENDIAN),
+      minY: header.getFloat64(44, LITTLE_ENDIAN),
+      minZ: header.getFloat64(68, LITTLE_ENDIAN),
+      minM: header.getFloat64(84, LITTLE_ENDIAN),
+      maxX: header.getFloat64(52, LITTLE_ENDIAN),
+      maxY: header.getFloat64(60, LITTLE_ENDIAN),
+      maxZ: header.getFloat64(76, LITTLE_ENDIAN),
+      maxM: header.getFloat64(92, LITTLE_ENDIAN)
+    }
   };
 }
